first screen files commit

+++ Software/projet-f103/appli/PIN.c
@@ -2,7 +2,7 @@
  * PIN.c
  *
  *  Created on: 8 janv. 2022
- *      Author: porta
+ *      Author: Louison
  */
 
 #include "PIN.h"
@@ -11,18 +11,18 @@
 #define DEFAULT_PIN_CODE {0, 0, 0, 0}
 
 
-static int* pins[PIN_MAX_NUMBER];
+static int8_t* pins[PIN_MAX_NUMBER];
 
 
 
 
 void pinInit(void){
-	int default_pin_code[] = DEFAULT_PIN_CODE;
+	int8_t default_pin_code[] = DEFAULT_PIN_CODE;
 	pinAdd(default_pin_code);
 
 }
 
-bool_e pinAdd(int * newPin){
+bool_e pinAdd(int8_t * newPin){
 	uint8_t i;
 	for(i = 0; i<PIN_MAX_NUMBER; i++)
 		{
@@ -36,7 +36,7 @@ bool_e pinAdd(int * newPin){
 
 }
 
-bool_e pinUse(int * pin){
+bool_e pinUse(int8_t * pin){
 	uint8_t i;
 	for(i = 0; i<PIN_MAX_NUMBER; i++)
 		{
